Fail reference_files check only when boolean get_subs = false
Fail reference files check only when get_subs = false. If get_subs = true, it seems program flow should continue so the code can subsequently make the attempt to download the subs from opensubtitles.com. As is, the program halts because no subtitles exist in directory path ".mkv-episode-matcher\cache\data", but I want it to download the subtitles for me. The workaround is to manually download the subtitles each time and unpack them, rename them, etc., into this folder, which is quite labor intensive.